test: check against too large block size in cbor

+++ test/unit-test/utest_helpers.c
@@ -40,7 +40,7 @@
 
 CborError createOtaStreamingMessage( uint8_t * pMessageBuffer,
                                      size_t messageBufferSize,
-                                     int blockIndex,
+                                     int64_t blockIndex,
                                      uint8_t * pBlockPayload,
                                      size_t blockPayloadSize,
                                      size_t * pEncodedSize,
